qetl-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "qetl-sdk"
dynamic = ["version"]
description = "Python SDK for QETL Quantum Processing Pipeline"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "QETL Development Team", email = "dev@qetl.io"}
]
keywords = ["quantum", "bioinformatics", "pipeline", "sdk", "qetl"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.25.0",
    "pydantic>=1.8.0",
    "click>=8.0.0",
    "rich>=10.0.0",
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "asyncio-throttle>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "pytest-asyncio>=0.15.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "pre-commit>=2.15.0",
    "tox>=3.24.0",
]
jupyter = [
    "jupyter>=1.0.0",
    "ipywidgets>=7.6.0",
    "matplotlib>=3.4.0",
    "plotly>=5.0.0",
]
cloud = [
    "boto3>=1.18.0",
    "azure-storage-blob>=12.8.0",
    "google-cloud-storage>=1.42.0",
]
viz = [
    "matplotlib>=3.4.0",
    "plotly>=5.0.0",
    "seaborn>=0.11.0",
]
all = [
    "qetl-sdk[dev,jupyter,cloud,viz]"
]

[project.urls]
Homepage = "https://github.com/qetl/qetl-sdk"
Documentation = "https://qetl-sdk.readthedocs.io"
Repository = "https://github.com/qetl/qetl-sdk.git"
"Bug Tracker" = "https://github.com/qetl/qetl-sdk/issues"

[project.scripts]
qetl = "qetl_sdk.cli:main"

[tool.setuptools]
packages = ["qetl_sdk"]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "qetl_sdk.__version__"}

[tool.setuptools.package-data]
qetl_sdk = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["qetl_sdk"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["qetl_sdk"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
